Aggregate error types with union.
This isn't totally in line with Hindley–Milner type system, but neither
is TypeScript. The idea here is to concatenate the error types when
chaining Result with `flatMap`.

The typing still checks out for the case where you don't alter the error
type.

The DX improvement here is that it enables the type system to ensure
exhausitivity without having to annotate the call-site.